text_to_sql_agents/Alpha-SQL/requirements.txt
datasketch
nltk
chardet
python-dotenv
openai
pandas
loguru
tqdm
pyyaml
langchain
langchain-openai
weave
sqlglot[rs]